some fixes on the installation scrirpts

+++ scripts/single_script/ntl_install.sh
@@ -12,13 +12,12 @@ wget http://shoup.net/ntl/ntl-$VERSION.tar.gz
 tar -xzf ntl-$VERSION.tar.gz
 cd ntl-$VERSION/src
 
-echo $PREFIX
-LD_LIBRARY_PATH=$PREFIX/gmp-latest/lib:$LD_LIBRARY_PATH
-export LD_LIBRARY_PATH
+export LD_LIBRARY_PATH=$PREFIX/gmp-latest/lib:$PREFIX/libtool-latest/lib:$PREFIX/gcc-latest/lib64:$PREFIX/mpfr-latest/lib:$PREFIX/mpc-latest/lib:$PREFIX/ppl-latest/lib:$PREFIX/cloog-gcc-latest/lib:$LD_LIBRARY_PATH
+# export PATH=$PREFIX/libtool-latest/bin:$PREFIX/gcc-latest/bin:$PATH 
 
 echo "#### Building ntl library ####"
-./configure PREFIX=$PREFIX/ntl-$VERSION NTL_GMP_LIP=on SHARED=on GMP_PREFIX=$PREFIX/gmp-latest/
-make -j $SLOTS
+./configure CC=$CC CXX=$CXX CFLAGS="-mtune=native -O3" CXXFLAGS="-mtune=native -O3" LIBTOOL="$PREFIX/libtool-latest/bin/libtool --tag=CXX " PREFIX=$PREFIX/ntl-$VERSION NTL_GMP_LIP=on SHARED=on GMP_PREFIX=$PREFIX/gmp-latest/
+make -j$SLOTS
 
 # Check for failure
 RET=$?
@@ -27,7 +26,13 @@ if [ $RET -ne 0 ]; then
 fi
 
 echo "#### Installing ntl library ####"
-make PREFIX=$PREFIX/ntl-$VERSION install
+make PREFIX=$PREFIX/ntl-$VERSION NTL_GMP_LIP=on SHARED=on install
+
+# Check for failure
+RET=$?
+if [ $RET -ne 0 ]; then
+	exit $RET
+fi
 
 rm $PREFIX/ntl-latest
 ln -s $PREFIX/ntl-$VERSION $PREFIX/ntl-latest
